// src/app/models/OperacionVolquete.ts
export interface OperacionVolquete {
  labor_inicio: string;
  ubicacion_destino_id: number;
  ubicacion_destino: string;
  n_cucharas: number;
  material: string;
  observaciones: string;
  mineral: string;
  desmonte: string;
  relleno: string;
  numero_volquete: string;
  relave: string;
  scoop: string;
  total_viajes: number;
}